Backported refactored names and AVR legacies. Minor ELRSV2 improvement
